Change the class name of number to operand
Change each operator class name to operator

Script.js
Change the forEach loop to for loop

For loop provides more flexibility in targeting
the specific node that is to be configured.

updateDisplay() is added

Allow synchronus display update.

setOperand() is added

Initially, the forEach loop was used to add operand
on click to the display. But, it affects readability
since all codes were used in that particular block.

The display will show 0 on startup, to prevent user
from entering doubling zeros, a conditional statement
was set. By operation of the calculator, user is
required to indicate the first operand before the
operator can be set. Therefore, on entering one of
the operators, the firstOperand will be stored.

Fixed displayed of the secondOperand

When the firstOperand and firstOperator is stored,
the display will be empty on pressing a number. To
fix this, a conditional statement, firstOperand === display
is set. But, there's a limitation to it, when user
inputs the same number as the firstOperand, eg:
firstOperand = 123 and secondOperand = 1234, only 4
will be displayed on screen. So, count == 0 was
added and increment it when that statement executes.

To prevent trailing zeros from being inputted,
when display == '0', the display will still be
the operand inputted.

setOperator() is added

Calculator operation:
1 - operand          | 1. operator
2 - operator         | 2. operand
3 - operand          | 3. equal / operator
4 - equal / operator |

The firstOperator will be stored when one of the
operators is clicked. User can change their mind,
so the operator could be changed.

When the secondOperand is stored, and equal or
operator is clicked, the operate function will
be executed. Assign count to 0 in case the
value of operate = firstOperand.