Added equalSign()
On inserting a + b = , the calculator will execute
the calculation based on the operator selected. In
this case, the existing operand and operator must
all be assigned null since user are allowed to enter
a new operation. The firstOperator and secondOperand
used were stored in a temporary variable. If user
chose to press '=' again, the temp variable along
with the current displayed value will be used to
execute the operation.

If user executes '=' twice, the display will be set
to 0. This allows the firstOperand to be inputted
correctly without executing the addition assignment.